ch8: Memory footprint of internal structs compacted
Reverted to using uint32_t instead of uint64_t for most of the internal
unsigned int representations. uint64_t was chosen earlier due to
misunderstanding of a warning around the use of fseek() and its offset
parameter.
The smaller int type size also reducing the footprint of pages when
stored in memory and on disk as the larger range will not be necessary.

+++ include/cursor.h
@@ -16,7 +16,7 @@
  */
 typedef struct {
     Table *table;
-    uint64_t row_num;
+    uint32_t row_num;
     bool end_of_table;
 } Cursor;
